+++ tests/src/transaction.test.ts
@@ -8,13 +8,16 @@ import {
   TransactionInstruction,
   PublicKey,
   ComputeBudgetProgram,
+  AddressLookupTableProgram,
 } from "@solana/web3.js";
 import {
   airdropIfRequired,
   confirmTransaction,
+  CreateLookupTable,
   getSimulationComputeUnits,
   prepareTransactionWithCompute,
   sendTransaction,
+  sendVersionedTransaction,
 } from "../../src";
 import { sendAndConfirmTransaction } from "@solana/web3.js";
 import assert from "node:assert";
@@ -35,7 +38,6 @@ describe("confirmTransaction", () => {
       lamportsToAirdrop,
       1 * LAMPORTS_PER_SOL,
     );
-
     const signature = await sendAndConfirmTransaction(
       connection,
       new Transaction().add(
@@ -47,7 +49,6 @@ describe("confirmTransaction", () => {
       ),
       [sender],
     );
-
     await confirmTransaction(connection, signature);
   });
 });
@@ -63,36 +64,30 @@ describe("getSimulationComputeUnits", () => {
       1 * LAMPORTS_PER_SOL,
     );
     const recipient = Keypair.generate().publicKey;
-
     const sendSol = SystemProgram.transfer({
       fromPubkey: sender.publicKey,
       toPubkey: recipient,
       lamports: 1_000_000,
     });
-
     const sayThanks = new TransactionInstruction({
       keys: [],
       programId: MEMO_PROGRAM_ID,
       data: Buffer.from("thanks"),
     });
-
     const computeUnitsSendSol = await getSimulationComputeUnits(
       connection,
       [sendSol],
       sender.publicKey,
       [],
     );
-
     // TODO: it would be useful to have a breakdown of exactly how 300 CUs is calculated
     assert.equal(computeUnitsSendSol, 300);
-
     const computeUnitsSendSolAndSayThanks = await getSimulationComputeUnits(
       connection,
       [sendSol, sayThanks],
       sender.publicKey,
       [],
     );
-
     // TODO: it would be useful to have a breakdown of exactly how 3888 CUs is calculated
     // also worth reviewing why memo program seems to use so many CUs.
     assert.equal(computeUnitsSendSolAndSayThanks, 3888);
@@ -195,6 +190,109 @@ describe("Transaction utilities", () => {
     });
   });
 
+  test("send versioned transaction", async () => {
+    const connection = new Connection(LOCALHOST);
+    const sender = Keypair.generate();
+    await airdropIfRequired(
+      connection,
+      sender.publicKey,
+      2 * LAMPORTS_PER_SOL,
+      1 * LAMPORTS_PER_SOL,
+    );
+    const recipient = Keypair.generate();
+
+    const instructions = [
+      SystemProgram.transfer({
+        fromPubkey: sender.publicKey,
+        toPubkey: recipient.publicKey,
+        lamports: LAMPORTS_PER_SOL * 0.1,
+      }),
+    ];
+
+    const statusUpdates: any[] = [];
+    const signature = await sendVersionedTransaction(
+      connection,
+      instructions,
+      [sender],
+      1000,
+      [],
+      {
+        computeUnitBuffer: { multiplier: 1.1 },
+        onStatusUpdate: (status) => {
+          statusUpdates.push(status);
+          console.log("status", status);
+        },
+      },
+    );
+
+    assert.ok(signature);
+    assert.deepEqual(
+      statusUpdates.map((s) => s.status),
+      ["created", "signed", "sent", "confirmed"],
+    );
+
+    console.log("Versioned transaction signature", signature);
+  });
+
+  test("send versioned transaction with lookup table", async () => {
+    const connection = new Connection(LOCALHOST);
+    const sender = Keypair.generate();
+    await airdropIfRequired(
+      connection,
+      sender.publicKey,
+      2 * LAMPORTS_PER_SOL,
+      1 * LAMPORTS_PER_SOL,
+    );
+
+    const recipient = Keypair.generate();
+    const recipient2 = Keypair.generate();
+    const recipient3 = Keypair.generate();
+
+    const [lookupTableAddress, lookupTableAccount] = await CreateLookupTable(
+      connection,
+      sender,
+      [
+        sender.publicKey,
+        recipient.publicKey,
+        recipient2.publicKey,
+        recipient3.publicKey
+      ],
+    );
+
+    const instructions = [
+      SystemProgram.transfer({
+        fromPubkey: sender.publicKey,
+        toPubkey: recipient.publicKey,
+        lamports: LAMPORTS_PER_SOL * 0.1,
+      }),
+      SystemProgram.transfer({
+        fromPubkey: sender.publicKey,
+        toPubkey: recipient2.publicKey,
+        lamports: LAMPORTS_PER_SOL * 0.1,
+      }),
+      SystemProgram.transfer({
+        fromPubkey: sender.publicKey,
+        toPubkey: recipient3.publicKey,
+        lamports: LAMPORTS_PER_SOL * 0.1,
+      }),
+    ];
+
+    const signature = await sendVersionedTransaction(
+      connection,
+      instructions,
+      [sender],
+      10000,
+      [lookupTableAccount],
+    );
+
+    assert.ok(signature);
+
+    console.log(
+      "Versioned transaction with lookup tables signature",
+      signature,
+    );
+  });
+
   test("sendTransaction should prepare and send a transaction and add priority fee instructions", async () => {
     const connection = new Connection(LOCALHOST);
     const sender = Keypair.generate();
@@ -219,7 +317,7 @@ describe("Transaction utilities", () => {
       connection,
       transaction,
       [sender],
-      1000,
+      10000,
       {
         computeUnitBuffer: { multiplier: 1.1 },
         onStatusUpdate: (status) => {